+++ packages/web/src/views/swap/components/swap-liquidity/SwapLiquidity.stories.tsx
@@ -1,8 +1,8 @@
 import React from "react";
-import { ComponentStory, ComponentMeta } from "@storybook/react";
-import SwapLiquidity from "./SwapLiquidity";
-import { dummyLiquidityList } from "@views/swap/containers/swap-liquidity-container/SwapLiquidityContainer";
 import { css } from "@emotion/react";
+import { ComponentStory, ComponentMeta } from "@storybook/react";
+
+import SwapLiquidity, { dummyLiquidityList } from "./SwapLiquidity";
 
 export default {
   title: "swap/SwapLiquidity",
