+++ packages/web/src/views/token-detail/components/trending-crypto-card-list/trending-crypto-card/TrendingCryptoCard.stories.tsx
@@ -1,8 +1,9 @@
 import React from "react";
 import { ComponentStory, ComponentMeta } from "@storybook/react";
 
+import { MATH_NEGATIVE_TYPE } from "@constants/option.constant";
+
 import TrendingCryptoCard from "./TrendingCryptoCard";
-import { trendingCryptoListInit } from "@views/token-detail/containers/trending-crypto-card-list-container/TrendingCryptoCardListContainer";
 
 export default {
   title: "token/TrendingCryptoCard",
@@ -15,5 +16,16 @@ const Template: ComponentStory<typeof TrendingCryptoCard> = args => (
 
 export const Default = Template.bind({});
 Default.args = {
-  item: trendingCryptoListInit[0],
+  item: {
+    path: "1",
+    name: "HEX",
+    symbol: "HEX",
+    logoURI:
+      "https://raw.githubusercontent.com/Uniswap/assets/master/blockchains/ethereum/assets/0x2b591e99afE9f32eAA6214f7B7629768c40Eeb39/logo.png",
+    price: "$12,908.25",
+    change: {
+      status: MATH_NEGATIVE_TYPE.NEGATIVE,
+      value: "-17.43%",
+    },
+  },
 };
