+++ packages/web/src/views/token-detail/containers/trending-crypto-card-list-container/TrendingCryptoCardListContainer.tsx
@@ -15,40 +15,6 @@ import { ITrending } from "@repositories/token";
 import { formatPrice, formatRate } from "@utils/new-number-utils";
 
 import TrendingCryptoCardList from "../../components/trending-crypto-card-list/TrendingCryptoCardList";
-import { TrendingCryptoInfo } from "../../components/trending-crypto-card-list/trending-crypto-card/TrendingCryptoCard";
-
-const trendingCryptoInit: TrendingCryptoInfo[] = [
-  {
-    path: "1",
-    name: "HEX",
-    symbol: "HEX",
-    logoURI:
-      "https://raw.githubusercontent.com/Uniswap/assets/master/blockchains/ethereum/assets/0x2b591e99afE9f32eAA6214f7B7629768c40Eeb39/logo.png",
-    price: "$12,908.25",
-    change: {
-      status: MATH_NEGATIVE_TYPE.NEGATIVE,
-      value: "-17.43%",
-    },
-  },
-  {
-    path: "2",
-    name: "USDCoin",
-    symbol: "USDC",
-    logoURI:
-      "https://raw.githubusercontent.com/Uniswap/assets/master/blockchains/ethereum/assets/0xA0b86991c6218b36c1d19D4a2e9Eb0cE3606eB48/logo.png",
-    price: "$12,908.25",
-    change: {
-      status: MATH_NEGATIVE_TYPE.POSITIVE,
-      value: "+17.43%",
-    },
-  },
-];
-
-export const trendingCryptoListInit = [
-  ...trendingCryptoInit,
-  ...trendingCryptoInit,
-  trendingCryptoInit[0],
-];
 
 const TrendingCryptoCardListContainer: React.FC = () => {
   const router = useCustomRouter();
